requirements.txt
Django==2.2
django-cors-headers==3.5.0
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
mysqlclient==2.0.1
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.2.4
